Move eliminate_bool_vectors to a header.
