+++ Makefile
@@ -283,6 +283,7 @@ SOURCE_FILES = \
   DeviceArgument.cpp \
   DeviceInterface.cpp \
   EarlyFree.cpp \
+  EliminateBoolVectors.cpp \
   Error.cpp \
   FastIntegerDivide.cpp \
   FindCalls.cpp \
@@ -403,6 +404,7 @@ HEADER_FILES = \
   DeviceArgument.h \
   DeviceInterface.h \
   EarlyFree.h \
+  EliminateBoolVectors.h \
   Error.h \
   Expr.h \
   ExprUsesVar.h \
